# apps/ios/Packages/DenAPI/Sources/DenAPI/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Den API",
    "description": "M3 REST API. Cookie-authenticated writes require Origin and X-CSRF-Token. CLI and agents use bearer authentication. Entity IDs are ULIDs; background image IDs are opaque SHA-256 content IDs.",
    "license": {
      "name": "AGPL-3.0-only",
      "identifier": "AGPL-3.0-only"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/channels/{id}/objects": {
      "post": {
        "tags": [
          "objects"
        ],
        "operationId": "post__channels__id__objects",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateObject"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Object"
                }
              }
            }
          },
          "default": {
            "description": "API error; invalid input, authentication, permissions, conflict, throttling or storage failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/instance": {
      "get": {
        "tags": [
          "objects"
        ],
        "operationId": "get__instance",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Instance"
                }
              }
            }
          },
          "default": {
            "description": "API error; invalid input, authentication, permissions, conflict, throttling or storage failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": []
      }
    },
...
    "/objects/{id}": {
      "get": {
        "tags": [
          "objects"
        ],
        "operationId": "get__objects__id_",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Object"
                }
              }
            }
          },
          "default": {
            "description": "API error; invalid input, authentication, permissions, conflict, throttling or storage failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "objects"
        ],
        "operationId": "patch__objects__id_",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateObject"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectSummary"
                }
              }
            }
          },
          "default": {
            "description": "API error; invalid input, authentication, permissions, conflict, throttling or storage failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/objects/{id}/patch": {
      "post": {
        "tags": [
          "objects"
        ],
        "operationId": "post__objects__id__patch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ObjectPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectVersion"
                }
              }
            }
          },
          "default": {
            "description": "API error; invalid input, authentication, permissions, conflict, throttling or storage failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/objects/{id}/summary": {
      "get": {
        "tags": [
          "objects"
        ],
        "operationId": "get__objects__id__summary",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectSummary"
                }
              }
            }
          },
          "default": {
            "description": "API error; invalid input, authentication, permissions, conflict, throttling or storage failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/settings": {
      "get": {
        "tags": [
          "objects"
        ],
        "operationId": "get__settings",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Settings"
                }
              }
            }
          },
          "default": {
            "description": "API error; invalid input, authentication, permissions, conflict, throttling or storage failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "objects"
        ],
        "operationId": "put__settings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSettings"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Settings"
                }
              }
            }
          },
          "default": {
            "description": "API error; invalid input, authentication, permissions, conflict, throttling or storage failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/uploads/{id}/file": {
      "get": {
        "tags": [
          "uploads"
        ],
        "operationId": "get__uploads__id__file",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Range",
            "in": "header",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "File bytes"
          },
          "206": {
            "description": "Requested byte range"
          },
          "416": {
            "description": "Unsatisfiable range"
          },
          "default": {
            "description": "API error; invalid input, authentication, permissions, conflict, throttling or storage failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "head": {
        "tags": [
          "uploads"
        ],
        "operationId": "file_head",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Range",
            "in": "header",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "File bytes"
          },
          "206": {
            "description": "Requested byte range"
          },
          "416": {
            "description": "Unsatisfiable range"
          },
          "default": {
            "description": "API error; invalid input, authentication, permissions, conflict, throttling or storage failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "CreateObject": {
        "type": "object",
        "required": [
          "kind",
          "name"
        ],
        "properties": {
          "kind": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "reply_to": {
            "type": [
              "string",
              "null"
            ]
          },
          "state": {
            "type": "object",
            "additionalProperties": {},
            "propertyNames": {
              "type": "string"
            }
          },
          "task_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "thread_id": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "Instance": {
        "type": "object",
        "required": [
          "instance_name",
          "version"
        ],
        "properties": {
          "icon_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "instance_name": {
            "type": "string"
          },
          "version": {
            "type": "string"
          }
        }
      },
...
      "Object": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ObjectSummary"
          },
          {
            "type": "object",
            "required": [
              "state"
            ],
            "properties": {
              "state": {
                "type": "object",
                "additionalProperties": {},
                "propertyNames": {
                  "type": "string"
                }
              }
            }
          }
        ]
      },
      "ObjectPatch": {
        "type": "object",
        "required": [
          "base_version"
        ],
        "properties": {
          "base_version": {
            "type": "integer",
            "format": "int64"
          },
          "put": {
            "type": "array",
            "items": {}
          },
          "remove": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/String"
            }
          }
        }
      },
...
      "ObjectSummary": {
        "type": "object",
        "description": "Document records keyed by record id. The server treats record contents as opaque.",
        "required": [
          "id",
          "channel_id",
          "kind",
          "name",
          "version",
          "created_by",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "channel_id": {
            "$ref": "#/components/schemas/String"
          },
          "created_at": {
            "type": "string"
          },
          "created_by": {
            "$ref": "#/components/schemas/String"
          },
          "id": {
            "$ref": "#/components/schemas/String"
          },
          "kind": {
            "type": "string"
          },
          "message_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "thumbnail_upload_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "thumbnail_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "updated_at": {
            "type": "string"
          },
          "version": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ObjectVersion": {
        "type": "object",
        "required": [
          "version"
        ],
        "properties": {
          "version": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "Settings": {
        "type": "object",
        "required": [
          "canvas_enabled",
          "instance_name"
        ],
        "properties": {
          "canvas_enabled": {
            "type": "boolean"
          },
          "instance_name": {
            "type": "string"
          }
        }
      },
...
      "String": {
        "type": "string"
      },
...
      "UpdateObject": {
        "type": "object",
        "properties": {
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "thumbnail_upload_id": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "UpdateSettings": {
        "type": "object",
        "properties": {
          "canvas_enabled": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "instance_name": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "den_session"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    },
    {
      "sessionCookie": []
    }
  ]
}